Dockerize Build
This change adds a `Dockerfile` which uses multi-stage build directives
to build a minimal image of the `web-api-commander` jar file.

This will allow for easily reproducable builds across development
environments.

The docker image artifact can also be used in environments where it is not
conveient to install Java such as a shared Jenkins environment.